move styles into component directory
