+++ packages/spinner/src/chameleon-spinner-style.ts
