+++ packages/spinner/src/materialize-loader-style.ts
